backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-dotenv>=1.0.1
